+++ src/astro/nutation/nut00a.rs
@@ -35,9 +35,9 @@
 //! * Wallace & Capitaine (2006), *Astron. Astrophys.* 459, 981
 //! * SOFA routines `iauNut00a`, `iauNut06a`
 
+use super::nut00a_tables::{NUT00A_LS, NUT00A_PL};
 use super::NutationAngles;
 use crate::astro::precession::mean_obliquity_iau2006;
-use crate::data::compiled::nut00a_tables::{NUT00A_LS, NUT00A_PL};
 use crate::qtty::*;
 use crate::time::JulianDate;